1. working on being able to set power 2. added a good example from West's class for sensitivity analysis
I included an example of how to calculate power if we set the current
just for kicks.

Then I started with the issue of calculating current needed to meet a
selected power. My first idea was to use scipy.optimize.minimize
function because it is the simplest I could think of, where it would
minimize power - I*V, but I think it found a local minimum or the
negatives got messed up. Will look into this!

Next, with my random power / concentration values, I made a widget so we
could see the shape of the Power vs current response.